+++ include/a3d/rendering/context/platform/desktop/Window.h
@@ -55,8 +55,8 @@ namespace a3d {
 
 		Window(RenderingApi renderAPI,
 			   std::string title,
-			   size_t width,
-			   size_t height,
+			   int width,
+			   int height,
 			   bool fullScreen,
 			   bool useHighDPI = true,
 			   AntialiasingMode antialiasingMode = AntialiasingMode::None);
